api: move `get_proc_address` to `GlDisplay`
This should made it clear that you should call this only once
and not on each context creation.
