+++ glutin/src/api/egl/context.rs
@@ -1,6 +1,5 @@
 //! Everything related to `EGLContext` management.
 
-use std::ffi::{self, CStr};
 use std::fmt;
 use std::marker::PhantomData;
 use std::ops::Deref;
@@ -227,10 +226,6 @@ impl PossiblyCurrentGlContext for PossiblyCurrentContext {
             self.inner.bind_api();
             self.inner.display.inner.egl.GetCurrentContext() == *self.inner.raw
         }
-    }
-
-    fn get_proc_address(&self, addr: &CStr) -> *const ffi::c_void {
-        unsafe { self.inner.display.inner.egl.GetProcAddress(addr.as_ptr()) as *const _ }
     }
 }
