+++ glutin/src/api/egl/display.rs
@@ -1,7 +1,7 @@
 //! Everything related to `EGLDisplay`.
 
 use std::collections::HashSet;
-use std::ffi::CStr;
+use std::ffi::{self, CStr};
 use std::fmt;
 use std::ops::Deref;
 use std::sync::Arc;
@@ -270,6 +270,10 @@ impl GlDisplay for Display {
         surface_attributes: &SurfaceAttributes<PixmapSurface>,
     ) -> Result<Self::PixmapSurface> {
         unsafe { Self::create_pixmap_surface(self, config, surface_attributes) }
+    }
+
+    fn get_proc_address(&self, addr: &CStr) -> *const ffi::c_void {
+        unsafe { self.inner.egl.GetProcAddress(addr.as_ptr()) as *const _ }
     }
 }
